+++ scg-impl/src/test/java/com/dwarfeng/scg/impl/configuration/FastJsonConfiguration.java
@@ -1,7 +1,10 @@
 package com.dwarfeng.scg.impl.configuration;
 
 import com.alibaba.fastjson.parser.ParserConfig;
-import com.dwarfeng.scg.sdk.bean.entity.*;
+import com.dwarfeng.scg.sdk.bean.entity.FastJsonCommonVariable;
+import com.dwarfeng.scg.sdk.bean.entity.FastJsonGeneratorSupport;
+import com.dwarfeng.scg.sdk.bean.entity.FastJsonNodeVariable;
+import com.dwarfeng.scg.sdk.bean.entity.FastJsonScgSetting;
 import com.dwarfeng.scg.sdk.bean.key.FastJsonScgNodeKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,7 +19,6 @@ public class FastJsonConfiguration {
         LOGGER.info("正在配置 FastJson auto-type 白名单...");
         ParserConfig.getGlobalInstance().addAccept(FastJsonGeneratorSupport.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonScgSetting.class.getCanonicalName());
-        ParserConfig.getGlobalInstance().addAccept(FastJsonScgNodeInfo.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonScgNodeKey.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonNodeVariable.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonCommonVariable.class.getCanonicalName());